requirements.txt
Flask
python-docx
lxml
gunicorn
docx2pdf; sys_platform == 'win32'
pywin32; sys_platform == 'win32'
